Implement BizfiBankConnection
Implemented BizfiBankConnection.GetAccountDetails()

+++ src/BankApi.Logic/BankConnections/Data/AccountDetails.cs
@@ -5,7 +5,7 @@
         public string AccountName { get; set; }
         public string AccountNumber { get; set; }
         public string SortCode { get; set; }
-        public decimal CurrentBalance { get; set; }
-        public decimal OverdraftLimit { get; set; }
+        public double CurrentBalance { get; set; }
+        public double OverdraftLimit { get; set; }
     }
 }
